+++ src/me/anno/engine/ui/control/DCDroppable.kt
@@ -5,12 +5,14 @@ import me.anno.ecs.Entity
 import me.anno.ecs.prefab.Prefab
 import me.anno.ecs.prefab.PrefabSaveable
 import org.joml.Quaterniond
+import org.joml.Quaternionf
 import org.joml.Vector3d
+import org.joml.Vector3f
 
 interface DCDroppable {
     fun drop(
         self: DraggingControls, prefab: Prefab, hovEntity: Entity?, hovComponent: Component?,
-        dropPosition: Vector3d, dropRotation: Quaterniond, dropScale: Vector3d,
+        dropPosition: Vector3d, dropRotation: Quaternionf, dropScale: Vector3f,
         results: MutableCollection<PrefabSaveable>
     )
 }
